user and post association done

+++ controllers/adminController.js
@@ -1,20 +1,14 @@
 const Post = require('../models/Post');
+const User = require('../models/User');
 
 const dashboard = (req, res, next) => {
-	// console.log(req.session.user);
-	// if(!req.session.user){
-	// console.log(req.session.user);
-		
-	// 	return res.redirect('/login');
-	// }
-
 	res.render('admin/dashboard', {
 		path:'/admin/dashboard'
 	}); 
 }
 
 const posts = (req, res) => {
-	Post.findAll().then(posts => {
+	Post.findAll({where: {userId: req.session.user.id}}).then(posts => {
 	 	res.render('admin/post/posts', {
 	 	all_post: posts,
 	 	path:'/admin/posts'
@@ -38,19 +32,27 @@ const storePost = (req, res) => {
 	}
 
 	const imageUrl = image.path;
+	const userId = req.session.user.id;
+	User.findByPk(userId)
+	.then(user => {
+		user.createPost({ title, imageUrl, description, author })
+		.then(post => {
+			req.flash('success_msg', 'Post created successfully.');
+			res.redirect('/admin/posts');
+		})
+		.catch(err => console.log(err));
 
-	Post.create({ title, imageUrl, description, author })
-	.then(post => {
-		req.flash('success_msg', 'Post created successfully.');
-		res.redirect('/admin/posts');
-	})
-	.catch(err => console.log(err));
+	});
 	
 }
 
 const editPost = (req, res) => {
-	Post.findOne( {where: {id: req.params.id}} )
+	Post.findOne( {where: {id: req.params.id, userId: req.session.user.id}} )
 	.then(data => {
+			if(!data){
+				req.flash('error_msg', 'You can only edit your own post.');
+				res.redirect('/admin/posts');
+			}
 			res.render('admin/post/edit-post', {
 			post: data,
 			path:'/admin/edit-post'
@@ -94,12 +96,15 @@ const updatPost = (req, res) => {
 }
 
 const deletePost = (req, res) => {
-	Post.destroy({where:{id:req.body.id}})
+	Post.destroy({where:{id:req.body.id, userId: req.session.user.id}})
 	.then(() => {
 		req.flash('success_msg', 'Post deleted successfully.');
 		res.redirect('/admin/posts');
 	})
-	.catch(err => console.log(err));
+	.catch(err => {
+		req.flash('error_msg', 'You can only delete your own post.');
+		res.redirect('/admin/posts');
+	});
 }
 
 module.exports = {dashboard, posts, addPost, storePost, editPost, updatPost, deletePost}
